[fix] Delete the minified file exits in source
This setup is wrong.  The data at this point is the same since 1.5.1 release.  Further, its check is incorrect.  Its looking to see if file is in source, well it should be checking target which would be nothing more than dest.exists() but since they are the same, just delete this for now.  The idea is good but not seeing how to achieve it as files seem to point to target based off classpath pretty early on since resources got processed by maven.