[ci] Cleanup how to build info and add a few more notes
original build information was total overkill.  That info might be useful for understanding how maven works in various ways but all of it built in the normal cycle and the code I added will release without concerns as the base-parent is mine and fully setup for doing all release items.